requirements.txt
torch>=2.2.0
torchvision>=0.17.0
transformers>=4.45.0
accelerate>=0.30.0
qwen-vl-utils>=0.0.8
flash-attn>=2.5.0
Pillow>=10.0.0
pandas>=2.0.0
pyyaml>=6.0
tqdm>=4.66.0
